Always generate FM at the mixer's output rate
To test this question:

  - Should we run the Adlib OPL emulator at 49716 Hz and
    then resample back down to the mixer's output rate or,

  - Should we run the Adlib OPL emulator at the mixer's
    output rate (eliminating the CPU cost of the higher OPL
    rate as well as the CPU cost, latency, and added arifacts
    of resampling).

Two recordings were performed using Blues Brothers:

 1.  oplrate = 49716 Hz resampled back down to 16000 Hz.
 2.  oplrate = 16000 Hz, saved natively at 16000 Hz.

Visually there was no difference in the wave forms (zoomed in or out
in Audacity). Likewise there was no audible difference when listening
to them with decent Sennheiser headphones, and repeating sections
over and over.

Perhaps the first recording still contains more information? To answer
this, FLAC compression was used to measure their information content
(or entropy), which from information theory says that the entropy rate
of a data source can be measured by the average number of bits per symbol
needed to encode it.

FLAC was able to compress both recordings to the same reported level of
compression and inspecting the actual byte-quantities in the resulting
FLAC files differed by a mere 0.039%.

So this answers it.. we're not losing any information by running OPL at
the mixer's direct output rate, and in doing so we avoid:

 1. wasting CPU cycles by over-sampling the OPL
 2. wasting CPU cycles by resampling the FM channel back down in the mixer
 3. introducing resampling artifacts in the output stream